Separate helper functions from Calendar component
